+++ lib/main.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:shop/screens/login/login_screen.dart';
-import 'package:shop/screens/login/signup_screen.dart';
+import 'package:shop/router.dart';
 import 'package:shop/theme/colors.dart';
 
 void main() {
@@ -22,26 +21,7 @@ class MyApp extends StatelessWidget {
     redirect: (context, state) {
       return null;
     },
-    routes: [
-      GoRoute(
-        path: '/signup',
-        builder: (context, state) => const SignUpScreen(),
-      ),
-      GoRoute(
-        path: '/login',
-        builder: (context, state) => const LoginScreen(),
-      ),
-      GoRoute(
-        path: '/product_card/:id',
-        builder: (context, state) {
-          final id = state.pathParameters['id'];
-          if (id == null) {
-            return const ErrorPage(error: 'Product ID is missing');
-          }
-          return Container();
-        },
-      ),
-    ],
+    routes: routes,
   );
 
   @override
@@ -56,20 +36,6 @@ class MyApp extends StatelessWidget {
       ),
       debugShowCheckedModeBanner: false,
       routerConfig: _router,
-    );
-  }
-}
-
-class ErrorPage extends StatelessWidget {
-  final String error;
-
-  const ErrorPage({super.key, required this.error});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Error')),
-      body: Center(child: Text('Error: $error')),
     );
   }
 }
